+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- Add model weights distribution tracking (gorarakelyan)
 - Add aim correlation tracking (gorarakelyan)
 
 ## 0.2.6 Nov 5 2019